+++ 1.20.4/Common/src/main/java/fuzs/easyanvils/config/ClientConfig.java
@@ -6,4 +6,6 @@ import fuzs.puzzleslib.api.config.v3.ConfigCore;
 public class ClientConfig implements ConfigCore {
     @Config(description = "Render inventory contents of an anvil.")
     public boolean renderAnvilContents = true;
+    @Config(description = "Add a tooltip to name tag items explaining how the change the name on the fly.")
+    public boolean nameTagTooltip = true;
 }