+++ 1.20.4/Common/src/main/java/fuzs/easyanvils/config/CommonConfig.java
@@ -1,0 +1,9 @@
+package fuzs.easyanvils.config;
+
+import fuzs.puzzleslib.api.config.v3.Config;
+import fuzs.puzzleslib.api.config.v3.ConfigCore;
+
+public class CommonConfig implements ConfigCore {
+    @Config(description = "Enable a crafting recipe for name tags.")
+    public boolean nameTagCraftingRecipe = false;
+}