+++ 1.20.4/Common/src/main/java/fuzs/easyanvils/handler/BlockConversionHandler.java
@@ -49,7 +49,7 @@ public class BlockConversionHandler {
     }
 
     public static EventResultHolder<InteractionResult> onUseBlock(Player player, Level level, InteractionHand interactionHand, BlockHitResult hitResult) {
-        if (!EasyAnvils.CONFIG.get(ServerConfig.class).disableVanillaAnvil) return EventResultHolder.pass();
+        if (!EasyAnvils.CONFIG.get(ServerConfig.class).miscellaneous.disableVanillaAnvil) return EventResultHolder.pass();
         BlockState blockState = level.getBlockState(hitResult.getBlockPos());
         if (BLOCK_CONVERSIONS.containsKey(blockState.getBlock()) && !blockState.is(ModRegistry.UNALTERED_ANVILS_BLOCK_TAG)) {
             player.displayClientMessage(Component.empty().append(INVALID_BLOCK_COMPONENT).withStyle(ChatFormatting.RED), true);