+++ 1.20.4/NeoForge/src/main/java/fuzs/easyanvils/neoforge/client/EasyAnvilsNeoForgeClient.java
@@ -2,7 +2,9 @@ package fuzs.easyanvils.neoforge.client;
 
 import fuzs.easyanvils.EasyAnvils;
 import fuzs.easyanvils.client.EasyAnvilsClient;
+import fuzs.easyanvils.data.client.ModLanguageProvider;
 import fuzs.puzzleslib.api.client.core.v1.ClientModConstructor;
+import fuzs.puzzleslib.neoforge.api.data.v2.core.DataProviderHelper;
 import net.neoforged.api.distmarker.Dist;
 import net.neoforged.bus.api.SubscribeEvent;
 import net.neoforged.fml.common.Mod;
@@ -14,5 +16,6 @@ public class EasyAnvilsNeoForgeClient {
     @SubscribeEvent
     public static void onConstructMod(final FMLConstructModEvent evt) {
         ClientModConstructor.construct(EasyAnvils.MOD_ID, EasyAnvilsClient::new);
+        DataProviderHelper.registerDataProviders(EasyAnvils.MOD_ID, ModLanguageProvider::new);
     }
 }